+++ Public/Start-LMSessionSyncServer.ps1
@@ -1,3 +1,32 @@
+<#
+.SYNOPSIS
+    This function starts a session synchronization server for Logic Monitor.
+
+.DESCRIPTION
+    The Start-LMSessionSyncServer function starts a Pode server that is used to synchronize sessions for Logic Monitor. 
+    It uses a secret vault to store session details. If the vault does not exist, it creates one. 
+    The function also provides options to enable request logging and error logging.
+
+.PARAMETER EnableRequestLogging
+    A switch to control whether request logging is enabled.
+
+.PARAMETER EnableErrorLogging
+    A switch to control whether error logging is enabled.
+
+.EXAMPLE
+    Start-LMSessionSyncServer -EnableRequestLogging -EnableErrorLogging
+
+    This command starts the session synchronization server with request logging and error logging enabled.
+
+.INPUTS
+    None. You cannot pipe objects to Start-LMSessionSyncServer.
+
+.OUTPUTS
+    The function does not return any output. It writes messages to the host to indicate the status of the server and the secret vault.
+
+.NOTES
+    The function throws an error if it fails to unlock the secret vault with the provided credentials.
+#>
 Function Start-LMSessionSyncServer {
     Param(
         [Switch]$EnableRequestLogging,